Explicitly cast `std::numbers::pi` into `float`.
This is to clarify that the loss of precision is intentional and
accepted.

+++ code/ajokki/ajokki_helsinki_east_downtown_scene.cpp
@@ -586,7 +586,7 @@ namespace ajokki
                 cat_object_struct1.local_name = "kissa" + index_string;
                 cat_object_struct1.brain = rest_brain;
                 cat_object_struct1.initial_rotate_vectors = { glm::vec3(0.0f, 1.0f, 1.0f), glm::vec3(0.0f, 0.0f, 1.0f) };
-                cat_object_struct1.initial_rotate_angles = { std::numbers::pi, -0.5f * std::numbers::pi };
+                cat_object_struct1.initial_rotate_angles = { static_cast<float>(std::numbers::pi), -0.5f * static_cast<float>(std::numbers::pi) };
                 cat_object_struct1.original_scale_vector = glm::vec3(first_cat_scale, first_cat_scale, first_cat_scale);
                 cat_object_struct1.cartesian_coordinates = glm::vec3(
                         first_cat_x + static_cast<float>(i) * translation_x,
@@ -611,7 +611,7 @@ namespace ajokki
             yli::ontology::HolobiontStruct turbo_polizei_png_holobiont_struct1(*helsinki_east_downtown_scene, *turbo_polizei_png_symbiosis);
             turbo_polizei_png_holobiont_struct1.brain = rest_brain;
             turbo_polizei_png_holobiont_struct1.initial_rotate_vectors = { glm::vec3(0.0f, 1.0f, 1.0f) };
-            turbo_polizei_png_holobiont_struct1.initial_rotate_angles = { std::numbers::pi };
+            turbo_polizei_png_holobiont_struct1.initial_rotate_angles = { static_cast<float>(std::numbers::pi) };
             turbo_polizei_png_holobiont_struct1.original_scale_vector = glm::vec3(1.0f, 1.0f, 1.0f);
             turbo_polizei_png_holobiont_struct1.cartesian_coordinates = glm::vec3(85.00f, -160.00f, 30.00f);
 
@@ -630,7 +630,7 @@ namespace ajokki
             yli::ontology::HolobiontStruct turbo_polizei_png_holobiont_struct2(*helsinki_east_downtown_scene, *turbo_polizei_png_symbiosis);
             turbo_polizei_png_holobiont_struct2.brain = rest_brain;
             turbo_polizei_png_holobiont_struct2.initial_rotate_vectors = { glm::vec3(0.0f, 1.0f, 1.0f) };
-            turbo_polizei_png_holobiont_struct2.initial_rotate_angles = { std::numbers::pi };
+            turbo_polizei_png_holobiont_struct2.initial_rotate_angles = { static_cast<float>(std::numbers::pi) };
             turbo_polizei_png_holobiont_struct2.original_scale_vector = glm::vec3(1.0f, 1.0f, 1.0f);
             turbo_polizei_png_holobiont_struct2.cartesian_coordinates = glm::vec3(85.00f, -175.00f, 30.00f);
